Refactorización de plugins
