+++ assets/AppAsset.php
@@ -24,10 +24,8 @@ class AppAsset extends AssetBundle
         'css/bootstrap-extra.css',
         'css/nav.css',
         'css/sidebar.css',
-        'css/bootstrap-colorselector.css',
     ];
     public $js = [
-        'js/bootstrap-colorselector.js',
         'js/jquery.cookie.js',
     ];
     public $depends = [
